[PM-3660] chore: move guid utils away from platform utils
