+++ libs/common/src/vault/services/fido2/fido2-authenticator.service.spec.ts
@@ -23,6 +23,7 @@ import { LoginView } from "../../models/view/login.view";
 
 import { AAGUID, Fido2AuthenticatorService } from "./fido2-authenticator.service";
 import { Fido2Utils } from "./fido2-utils";
+import { guidToRawFormat } from "./guid-utils";
 
 const RpId = "bitwarden.com";
 
@@ -112,7 +113,7 @@ describe("FidoAuthenticatorService", () => {
         params = await createParams({
           excludeCredentialDescriptorList: [
             {
-              id: Utils.guidToRawFormat(excludedCipher.login.fido2Key.nonDiscoverableId),
+              id: guidToRawFormat(excludedCipher.login.fido2Key.nonDiscoverableId),
               type: "public-key",
             },
           ],
@@ -186,7 +187,7 @@ describe("FidoAuthenticatorService", () => {
         excludedCipherView = createCipherView();
         params = await createParams({
           excludeCredentialDescriptorList: [
-            { id: Utils.guidToRawFormat(excludedCipherView.id), type: "public-key" },
+            { id: guidToRawFormat(excludedCipherView.id), type: "public-key" },
           ],
         });
         cipherService.get.mockImplementation(async (id) =>
@@ -633,7 +634,7 @@ describe("FidoAuthenticatorService", () => {
         credentialId = Utils.newGuid();
         params = await createParams({
           allowCredentialDescriptorList: [
-            { id: Utils.guidToRawFormat(credentialId), type: "public-key" },
+            { id: guidToRawFormat(credentialId), type: "public-key" },
           ],
           rpId: RpId,
         });
@@ -709,7 +710,7 @@ describe("FidoAuthenticatorService", () => {
         ];
         params = await createParams({
           allowCredentialDescriptorList: credentialIds.map((credentialId) => ({
-            id: Utils.guidToRawFormat(credentialId),
+            id: guidToRawFormat(credentialId),
             type: "public-key",
           })),
           rpId: RpId,
@@ -789,7 +790,7 @@ describe("FidoAuthenticatorService", () => {
             selectedCredentialId = credentialIds[0];
             params = await createParams({
               allowCredentialDescriptorList: credentialIds.map((credentialId) => ({
-                id: Utils.guidToRawFormat(credentialId),
+                id: guidToRawFormat(credentialId),
                 type: "public-key",
               })),
               rpId: RpId,
@@ -842,7 +843,7 @@ describe("FidoAuthenticatorService", () => {
           const flags = encAuthData.slice(32, 33);
           const counter = encAuthData.slice(33, 37);
 
-          expect(result.selectedCredential.id).toEqual(Utils.guidToRawFormat(selectedCredentialId));
+          expect(result.selectedCredential.id).toEqual(guidToRawFormat(selectedCredentialId));
           expect(result.selectedCredential.userHandle).toEqual(
             Fido2Utils.stringToBuffer(fido2Keys[0].userHandle)
           );
