+++ libs/common/src/vault/services/fido2/fido2-authenticator.service.ts
@@ -21,6 +21,7 @@ import { Fido2KeyView } from "../../models/view/fido2-key.view";
 
 import { joseToDer } from "./ecdsa-utils";
 import { Fido2Utils } from "./fido2-utils";
+import { guidToRawFormat, guidToStandardFormat } from "./guid-utils";
 
 // AAGUID: 6e8248d5-b479-40db-a3d8-11116f7e8349
 export const AAGUID = new Uint8Array([
@@ -185,7 +186,7 @@ export class Fido2AuthenticatorService implements Fido2AuthenticatorServiceAbstr
 
       const authData = await generateAuthData({
         rpId: params.rpEntity.id,
-        credentialId: Utils.guidToRawFormat(credentialId),
+        credentialId: guidToRawFormat(credentialId),
         counter: fido2Key.counter,
         userPresence: true,
         userVerification: userVerified,
@@ -200,7 +201,7 @@ export class Fido2AuthenticatorService implements Fido2AuthenticatorServiceAbstr
       );
 
       return {
-        credentialId: Utils.guidToRawFormat(credentialId),
+        credentialId: guidToRawFormat(credentialId),
         attestationObject,
         authData,
         publicKeyAlgorithm: -7,
@@ -295,7 +296,7 @@ export class Fido2AuthenticatorService implements Fido2AuthenticatorServiceAbstr
 
         const authenticatorData = await generateAuthData({
           rpId: selectedFido2Key.rpId,
-          credentialId: Utils.guidToRawFormat(selectedCredentialId),
+          credentialId: guidToRawFormat(selectedCredentialId),
           counter: selectedFido2Key.counter,
           userPresence: true,
           userVerification: userVerified,
@@ -310,7 +311,7 @@ export class Fido2AuthenticatorService implements Fido2AuthenticatorServiceAbstr
         return {
           authenticatorData,
           selectedCredential: {
-            id: Utils.guidToRawFormat(selectedCredentialId),
+            id: guidToRawFormat(selectedCredentialId),
             userHandle: Fido2Utils.stringToBuffer(selectedFido2Key.userHandle),
           },
           signature,
@@ -334,7 +335,7 @@ export class Fido2AuthenticatorService implements Fido2AuthenticatorServiceAbstr
 
     for (const credential of credentials) {
       try {
-        ids.push(Utils.guidToStandardFormat(credential.id));
+        ids.push(guidToStandardFormat(credential.id));
         // eslint-disable-next-line no-empty
       } catch {}
     }
@@ -365,7 +366,7 @@ export class Fido2AuthenticatorService implements Fido2AuthenticatorServiceAbstr
 
     for (const credential of credentials) {
       try {
-        ids.push(Utils.guidToStandardFormat(credential.id));
+        ids.push(guidToStandardFormat(credential.id));
         // eslint-disable-next-line no-empty
       } catch {}
     }
