+++ libs/common/src/vault/services/fido2/fido2-client.service.spec.ts
@@ -17,6 +17,7 @@ import {
 import { Fido2AuthenticatorService } from "./fido2-authenticator.service";
 import { Fido2ClientService } from "./fido2-client.service";
 import { Fido2Utils } from "./fido2-utils";
+import { guidToRawFormat } from "./guid-utils";
 
 const RpId = "bitwarden.com";
 
@@ -235,7 +236,7 @@ describe("FidoAuthenticatorService", () => {
 
     function createAuthenticatorMakeResult(): Fido2AuthenticatorMakeCredentialResult {
       return {
-        credentialId: Utils.guidToRawFormat(Utils.newGuid()),
+        credentialId: guidToRawFormat(Utils.newGuid()),
         attestationObject: randomBytes(128),
         authData: randomBytes(64),
         publicKeyAlgorithm: -7,
@@ -346,8 +347,8 @@ describe("FidoAuthenticatorService", () => {
     describe("assert non-discoverable credential", () => {
       it("should call authenticator.assertCredential", async () => {
         const allowedCredentialIds = [
-          Fido2Utils.bufferToString(Utils.guidToRawFormat(Utils.newGuid())),
-          Fido2Utils.bufferToString(Utils.guidToRawFormat(Utils.newGuid())),
+          Fido2Utils.bufferToString(guidToRawFormat(Utils.newGuid())),
+          Fido2Utils.bufferToString(guidToRawFormat(Utils.newGuid())),
           Fido2Utils.bufferToString(Utils.fromByteStringToArray("not-a-guid")),
         ];
         const params = createParams({
